add the infomation about the method of importabi
